Update patch set 1
Patch Set 1:

I'm concerned about committing to maintaining compatability with an old release that we don't test with, and which has API changes that affect how we use it.  For example, this appears to change from a single atomic create+get operation to two separate operations.  It's probably fine in this case, since we don't expect anyone to change the component information between these two calls, but in general, something like that is theoretically possible, and we would need to evaluate every such operation to determine whether it's okay.  Our use of ZK is pretty sophisticated and we spend a lot of time thinking about atomic operations and race conditions.  If I'm wrong about this case being okay, we won't detect the error.

Since this is the only place this shows up, and if this is the only such incompatability, I think I might be okay with this change, but realistically, I don't think we can promise <3.5 compatability, so if we do merge this, I don't want it to set a precedent.

How feasible is it for you to upgrade?

Patch-set: 1
CC: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>