Put zuul vars back in debug inventory.yaml
We pass zuul vars as extra vars, which means they don't need to be
in the inventory file.  But we also write a "debug" inventory file
which isn't actually used by Zuul, but is useful for examining the
build.  Go ahead and write zuul vars into that to maintain
compatability with the old behavior.

Change-Id: I06e1303beb507a45575a20264dd2071d58d8b39d