Add nodeset to build table
The name of the nodeset used by a job may be of interest to users
as they look at historical build results (did py38 run on focal or
bionic?).  Add a column for that purpose.

Meanwhile, the node_name column has not been used in a very long
time (at least v3).  It is universally null now.  As a singular value,
it doesn't make a lot of sense for a multinode system.  Drop it.

The build object "node_labels" field is also unused; drop it as well.

The MQTT schema is updated to match SQL.

Change-Id: Iae8444dfdd52561928c80448bc3e3158744e08e6