Update patch set 3
Patch Set 3: Code-Review+2

Mechanically this change seems fine. There was discussion on IRC about whether or not GHE supports the neutral status. If GHE does not then I still think this should work as long as users don't set neutral as the check status. In that case a small update to the docs to indicate this is probably all that is required to change.

Patch-set: 3
Reviewer: Gerrit User 4146 <4146@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2