wip: Items ahead/behind wrong after re-enqueue
Marked 'wip' as this is currently only adding a test case that
demonstrates the issue w/o fixing it.

After an item changes the queue during re-enqueue the item_ahead and
items_behind pointers are wrong.

This is not showing up as an error as the pipeline manager seems to
correct the wrong item state.

Change-Id: I943b147f4623704f627dfec2833ced852d51cbcc