Fix items ahead/behind wrong after re-enqueue
After an item changes the queue during re-enqueue the item_ahead and
items_behind pointers can be wrong. This happened because the pipeline
manager didn't check if the former item ahead is still in the same
queue.

This was not showing up as an error as the pipeline manager seems to
correct the wrong item state.

Change-Id: I943b147f4623704f627dfec2833ced852d51cbcc