Remove exception handling in executor api isLocked
This should not produce NoNodeError under any normal circumstance,
and so if it does produce an error, we should not intercept it
and let the calling site figure out how to deal with it.

The only call path for this method is through the periodic lost
request cleanup routine, which is started by apscheduler.  It will
log any exceptions (though if we would prefer to handle that ourselves,
we could do so).

Change-Id: Iaa5720b1112b3f09e0a30a0360dc68213868f454