Tenant reconfiguration: Allow ref-updated newrev+oldrev reconfiguration
2 different gerrit configurations have to be managed: ssh(+http) and http

stream-event is used: ssh(+http)
--------------------------------
ref-updated oldrev+newrev events now triggers reconfiguration if they are
generated from a direct push and not a change.
ref-updated link to a change are filtered and do not trigger tenant
reconfiguration.

1st ref-updated/2nd change-merged
When treating the ref-updated data, We inspect the event queue to look
for the associated change-merged. if found, branch_updated is not set,
which will not trigger a tenant reconfiguration.

1st change-merged/2nd ref-updated
The same idea could be used (ie: a lookup table to check when inspecting
ref-updated data, to look if the associated change-merged event was seen)
Instead a tenant/project/branch/revision table is used, see below.

stream-events is not used: http
-------------------------------
when stream-event is not used, change-merged events are filtered:
those do not provide the branch newRev but the latest change revision, so we
cannot link them to the ref-updated event that will be generated.
We also can't reliably fetch the current branch revision and link it to a
change. Having the proper newRev is mandatory for a following change.

The effect is that all ref-updated event, as they will have files to 'None'
will trigger a tenant reconfiguration.

tenant/project/branch/revision table
------------------------------------
A tenant/project/branch/revision table is created to avoid reconfiguring
multiple times. the table is filled in cacheLoadConfig at init time and
reconfiguration. If an event that modifies the configuration end up to an
already known state, there is no need to perform a reconfigure.
merger:cat is modified to provide commit revision alongside config files.

Change-Id: I67289eade655f4a690c3b4c92f94eeb11ebbbca5