Move tenant validation to separate method
In order to simplify and cleanup the reconfiguration event handling we
move to tenant validation to a separate method. This method will be
directly called from the scheduler cmd app, instead of handling the
validation via the reconfiguration event.

This should give us a clearer picture on the differences of smart and
full reconfigurations, so later on we might be able to have a single
method for handling the different types of reconfigs.

Change-Id: Ifb8715ea1436d1f7f3cc127a2be88d4f5f89e73d