Combine full and smart reconfiguration events
Full and smart reconfigurations only differ in a few areas, mainly which
tenans are reconfigured and clearing of the connection and unparsed
branch caches.

Clearing of connection caches doesn't seem to be needed in practice and
will be removed as this can't be easily coordinated between multiple
schedulers. However, since some tests seem to depend on that behavior
this will be done in a separate change.

We will also no longer create a new Abide instance for full
reconfigurations as this would also require additional coordination
between multiple schedulers if we wanted to replicate that.

Change-Id: I746d81547ad8e0e0f669981d0fcce65f9fd2e32b