Fix data watch race in test_cancel_starting_build
Through extra logging we found that the data watch event for the created
cancel node was firing just as the delete request was happening. In some
instances this could mean that the watch handler would try to read the
data and get back NoNodeError. When this happened kazoo saw the event as
a noop because the current state matched the initial non existing znode
state and it wouldn't call our data_watch callback. This prevented us
from setting the event that allows the test to proceed.

To fix this we use kazoo.client.exists() with a watch that will fire the
watch once as soon as the znode exists whether or not it gets
subsequently deleted. We don't actually care about the contents of the
znode, only that it existed at one time so this is sufficient for our
needs.

Change-Id: I537ec29dafe89c79f3df3ef16fdd5582d2f2e311