Fix data watch race in test_cancel_starting_build
Through extra logging we found that the data watch event for the created
cancel node was firing just as the delete request was happening. In some
instances this could mean that the watch handler would try to read the
data and get back NoNodeError. When this happened we never stopped the
watcher and set the event saying the cancel node had existed allowing
the test to continue.

To fix this in our data watcher we ignore the first event which is
called when we set the watcher. Then look for the second created event
regardless of its content. If we see that then we set the test event
indicating the znode has existed.

Change-Id: I537ec29dafe89c79f3df3ef16fdd5582d2f2e311