Update patch set 1
Patch Set 1:

Seems like the risk in making this change is that if we have schedulers that miss branch creation or deletion events we may need to restart them to see the up to date state. Over time these deltas may accumulate.

I'm also not sure why we need to coordinate these between schedulers. Seems like that as long as each of them is periodically clearing their caches that would be better than the proposed situation here where we could diverge forever? I think its ok if they diverge slightly from each other for short periods of time which is what we may get if we just clear caches periodically.

Patch-set: 1
CC: Gerrit User 4146 <4146@4a232e18-c5a9-48ee-94c0-e04e7cca6543>