Update patch set 3
Patch Set 3:

> Seems like the risk in making this change is that if we have schedulers that miss branch creation or deletion events we may need to restart them to see the up to date state. Over time these deltas may accumulate.

We've considered that and came to the conclusion that neither we nor OpenDev uses full reconfigurations a lot (which would clear the caches). It looks like this is not really an issue in production.

I'm not sure if it makes sense to maybe keep the clearCaches() API and introduce a scheduler cmd that allows us to perform this step manually if needed. Or we keep the API and rely on the REPL to clear the caches if needed (not sure though if this is something we want to encourage).

> I'm also not sure why we need to coordinate these between schedulers. Seems like that as long as each of them is periodically clearing their caches that would be better than the proposed situation here where we could diverge forever? I think its ok if they diverge slightly from each other for short periods of time which is what we may get if we just clear caches periodically.

I think periodically clearing the caches is a lot more risky than not clearing them at all. As explained above this seems to be much closer to what we do in production ATM.

Patch-set: 3