Update patch set 4
Patch Set 4:

Yeah, I think it's safe to remove this now since we never use it, and in the future if it becomes necessary, an operator can achieve the same effect with a rolling restart of schedulers.  And if that does happen, we could look into adding it back in.

Patch-set: 4
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>