Ensure config cache stages are used correctly
We have a local unparsed branch cache on each scheduler and the global
config cache in Zookeeper. In order to check if the local unparsed
branch cache is up-to-date, we need to also store the logical timestamp
similar to the Zookeeper config cache.

Depending on the event that triggers (re-)loading of a tenant we must
make sure that those caches are considered valid or invalid correctly.

Instead of clearing the cache for a project(-branch) we will use the
event ltime to detected an outdated project-branch cache. This way
parallel tenant reconfigurations will not all clear out the cache in
Zookeeper even though it might be valid again because of a previously
processed reconfiguration.

There are four scenarios to consider when loading a tenant.

1. Processing a tenant reconfig event:
   - The min. ltime for the changed project(-branches) will be set to
     the event's ``zuul_event_ltime`` (to establish a happened-before
     relation in respect to the config change).  The min. ltime for all
     other project-branches will be -1.
   - Config for needed project-branch(es) is updated via cat job if the
     cache is not valid (cache ltime < min. ltime).
   - Cache in Zookeeper and local unparsed branch cache is updated. The
     ltime of the cache will be the timestamp created shortly before
     requesting the config via the mergers (only for outdated items).
2. Processing a full reconfiguration event:
   - The min. ltime for all project-branches is given as the
     ``zuul_event_ltime`` of the reconfiguration event.
   - Config for needed project-branch(es) is updated via cat job if the
     cache is not valid (cache ltime < min. ltime).  Otherwise the local
     unparsed branch cache or the global config cache in Zookeper is
     used.
   - Cache in Zookeeper and local unparsed branch cache is updated, with
     the ltime shortly before requesting the config via the mergers
     (only for outdated items).
3. Processing a smart reconfiguration event:
   - The min. ltime for all project-branches is given as -1 in order to
     use cached data wherever possible.
   - Config for new project-branch(es) is updated via cat job if the
     project is not yet cached. Otherwise the local unparsed branch
     cache or the global config cache in Zookeper is used.
   - Cache in Zookeeper and local unparsed branch cache is updated, with
     the ltime shortly before requesting the config via the mergers
     (only for new items).
4. (Re-)loading a tenant due to a changed layout (happens after an event
   according to one of the other scenarios was processed on another
   scheduler):
   - The min. ltime for all project-branches is given as -1 in order to
     only use cached config.
   - Local unparsed branch cache is updated if needed.

Change-Id: I45ec6dd794413870f85e4f9e9de8122eddce222e

Use the ZK cache for smart reconfiguration events

Traditionally, a smart reconfiguration event has used the file cache
under the assumption that it's only adding or removing tenants, projects,
or branches.  The content of the file cache is expected to be correct.
A recent change (I6b7afd34203b74d5a6524e5277055a813147ff9e) altered
that so that the file cache for all project-branches in any tenant
that was updated by a smart reconfig is cleared.  This means than extra
cat jobs are dispatched in order to read files which should not have
actually changed.

This change restores the previous behavior by asserting that the ZK
file cache is usable for every project-branch in the case of a smart
reconfiguration.  A full reconfiguration will still invalidate the
entire ZK file cache.

Change-Id: I9c2951ef29e3388723c78d26555229168d81413c