Ensure config cache stages are used correctly
We have a local unparsed branch cache on each scheduler and the global
config cache in Zookeeper. In order to check if the local unparsed
branch cache is up-to-date, we need to also store the logical timestamp
similar to the Zookeeper config cache.

Depending on the event that requires (re-)loading a tenant we need to
make sure that the caches are considered valid or invalid correctly.
There are three scenarios to consider when loading a tenant:

1. processing a tenant reconfig event
   - the Zookeeper config cache for a particular project(-branch) is cleared
   - cache ltime is given as -1 in order to use cached config if
     possible (except for the cleared project(-branch))
   - config for needed project-branch(es) is updated via cat job
   - cache in Zookeeper and local unparsed branch cache is updated
2. processing a (smart) reconfiguration event
   - cache ltime is given as the ltime of the reconfig event
   - get config of project-branches via cat job if cache ltime < event ltime,
     otherwise use the content from the cache
   - cache in Zookeeper and local unparsed branch cache is updated
3. (re-)loading a tenant due to a changed layout (not yet implemented)
   - cache ltime is given as -1 in order to only use cached config
   - local unparsed branch cache is updated if outdated

Change-Id: I45ec6dd794413870f85e4f9e9de8122eddce222e