Improve performance of _saveRepoState
In case of large repository with more than 10k refs,
this method use actualy an async call from Gitpython to retrieve sha1.
Gitpython open file filesystem for each refs

For example with repository with 18k tags,
a merger instance take 100% of one CPU (not threadless) for ~ 3min
to perform the loop

To improve this, we store all sha1 of tag directly from
a git command (for_each_ref), this method open once the packedref
of repository to extract all refs.

If a ref is not in the dict we use fallback method `ref.object`

Also we have scope this new workflow to tags because
is normaly statics refs

Change-Id: I8b52b39cb79527791a34ac98a25e7ee41c8d4956