Use the ZK cache for smart reconfiguration events
Traditionally, a smart reconfiguration event has used the file cache
under the assumption that it's only adding or removing tenants, projects,
or branches.  The content of the file cache is expected to be correct.
A recent change (I6b7afd34203b74d5a6524e5277055a813147ff9e) altered
that so that the file cache for all project-branches in any tenant
that was updated by a smart reconfig is cleared.  This means than extra
cat jobs are dispatched in order to read files which should not have
actually changed.

This change restores the previous behavior by asserting that the ZK
file cache is usable for every project-branch in the case of a smart
reconfiguration.  A full reconfiguration will still invalidate the
entire ZK file cache.

Change-Id: I9c2951ef29e3388723c78d26555229168d81413c