Don't display in-progress builds/buildsets in the web listing
The main purpose of the build and buildset listings is to find
completed builds, not in-progress builds (that's the status page),
so when we query the Zuul API, always set "completed=true".

In the future, it might be nice to list in-progress builds on these
pages, but that will require significantly more UI work to add
it to the filter bar.

Additionally, update the buildset page to show in-progress builds.
It currently doesn't because it assumes all non-final builds are
retry attemtps and therefore stores them under a secondary key.
That's no longer true, and additionally, completed retried builds
now have an explicit "RETRY" result which makes them easy for
users to distinguish.

It's more useful for a user to be able to see that a buildset is
still in progress and also the retried builds (now that we store
all of them), so this change returns all the builds of a buildset.

Change-Id: I83d89c2903e6fd0aaf67c66f61d106b2c5db3797