Split ZK event queues by type
We're currently storing all of (trigger, result, management) events
in a single queue per tenant or per-pipeline.  This means that in
order to iterate over just the management events, we need to
filter the list by name.

By splitting the event queues, we will be able to deal with smaller
numbers of events at a time, and can more quickly calculate event
queue length.

Change-Id: I2c01e51a6f9c55ba521d26416db16dfba068246d