Sort tenant list by name
Since we no longer create a new abide instance during reconfigurations
and just replace the tenant instance, the tenants are not longer sorted
in the order they appear in the config but depending on which one was
recently reconfigured.

From a users perspective it makes sense that the list of tenants is
sorted alphabetically.

Change-Id: Ie9ae0700fd3928cc1f8f93d1de77537ce97518cd