Store runtime related system attributes together
This change introduces a new class for storing runtime related system
attributes that are coming from the Zuul config. The system attributes
will be updated when the config changes instead of directly reading them
from the config on demand.

This gives us a clear interface for updating those values and storing
them in Zookeeper later on.

System config attributes currently are:
* [scheduler].relative_priority
* [scheduler].max_hold_expiration
* [scheduler].default_hold_expiration
* [scheduler].default_ansible_version
* [web].root
* [web].status_url
* [web].websocket_url

Change-Id: Iefc7857eddc3e0c068b10d01489aa6ae75a6374b