Set runtime related system attributes on reconfig
Some runtime related attributes come from the Zuul config and should be
consistent across multiple schedulers. This means that they also need
to be stored in Zookeeper later on.

This change creates attributes for those config values and updates them
when the config changes instead of reading them from the config
on demand. Later, we will also update those values when the system
config in Zookeeper changes.

Those config attributes are:
* [scheduler].relative_priority
* [scheduler].max_hold_expiration
* [scheduler].default_hold_expiration
* [scheduler].default_ansible_version
* [web].root
* [web].status_url
* [web].websocket_url

Change-Id: Iefc7857eddc3e0c068b10d01489aa6ae75a6374b