Cache system config in Zookeeper
The Zuul system config consists of the unparsed abide config and runtime
related attributes from the Zuul config file. In order for this config
to be consistent on all active schedulers, we also need to store it in
Zookeeper.

The system config will be updated during processing of a reconfiguration
event. Other schedulers can detect the updated system config and perform
a local update.

Change-Id: I8aaf52cc66aa44a7eb166cc8e277d8499d7bb86a