Update patch set 6
Patch Set 6:

PS6 changes the merger_api.submit() parameters, so the MergeRequest is created internally. It also adds the event_id as attribute on the MergeRequest.

I've also removed the sharding of the merge results as this makes the management event results fail. A possible fix could be to shard those management results as well, but I didn't want to experiment too much on this change now. Maybe it's worth to keep the sharding as separate change and elaborate a little more on it.

Patch-set: 6