Web UI: Add "Create Autohold Request" form, improve API error messages
Add a "create autohold request" button on the autoholds page for admin
users.
Refactor autohold form modal so that it can be reused in autoholds and
build pages.
Improve error messages when using admin actions: if the API call returns
an error description, display it in the toast. Otherwise return the HTTP
error code.

Change-Id: I9d1cc8a1d751e7b4ce32e7abecc2ad9f3e96f676