wip: Support cross scheduler config loading
On startup a scheduler will use the cached system config from Zookeeper
in case it exists. Otherwise it will load/generate the config and save
it to the cache.

The scheduler will then load the configured tenants as follows:

- In case a layout state already exists the scheduler willl acquire the
  tenant READ lock, read the cached unparsed config from Zookeeper and
  load the tenant.
- If there is no existing layout state for a tenant the scheduler
  acquires the tenant WRITE lock, loads the tenant and creates a new
  layout state in Zookeeper.

This changes the behavior during config priming in that we no longer
perform a full reconfiguration on scheduler startup.

During processing of a tenant reconfiguration event, the layout state in
Zookeeper for that tenant is updated. All other schedulers are notified
of the changed layout state via a data watch, which will set the wake
event of the run handler. Similarliy, a changed system config in
Zookeeper will also wake up the run handler.

The system config and tenanty layouts are direcly updated in the run
handler when an outdated config/layout is detected.

Change-Id: I48b145cecec5890d212d751f68b3947f04ee1ed5