Update patch set 1
Patch Set 1:

Actually, it's not possible to move the executor cleanup into the api because it emits a result event, and the executor api doesn't have a result_events object.  Because of that, I'm not convinced moving it is a good idea.

I'll take one more look at whether we can test the merger cleanup code another way.

Patch-set: 1