Clean up leaked event queue side channel data
Since we can't use transactions, we need to clean up any leaked
side channel data nodes.  We do this at the end of processing each
event queue -- at this point, the queues are likely to have a low
number of nodes which should avoid us having to waste too much time
iterating over valid nodes.  It also avoids us having to have a
cleanup loop that iterates over all tenants/pipelines.

Since we must compare the contents of the side channel "directory"
with the contents of the queues, a global side channel root would
mean we have to examine the contents of every queue to decide if
it's valid.  That seems unecessarily complicated (especially since
queues can appear or disappear with reconfigurations), and would
necessitate complex locking.

Instead, we move the side channel data to reside next to the actual
queue data, so each queue has its own side channel "directory".
The actual queues are shifted down one level in the ZK hierarchy
so that the side channel data can share a ZK node root with them.

Change-Id: Ifbeb0acfe47732e7376a8aeae643bc7732e96513