Move executor api cleanup to apscheduler
We don't need to clean up leaked build parameters every time we
clean up lost build requests.  The latter happens every 60 seconds
because it's performance critical, but the param cleanup can happen
much less frequently.

Move it into the scheduled tasks that the scheduler runs, but rather
than continuing to add new apsched events, combine this with the
config cache cleanup.  We can also add the similar cleanup task which
will be needed by the merger api.  These can all run infrequently
(every 60 minutes).

The only downside to this is that we will have one scheduler running
all of these combined tasks, but they are relatively low impact so
it's probably not important to load-balance them.

Change-Id: I76ce9df242ebdaf9f1b7113530775bd702ccaaa0