Update patch set 4
Patch Set 4: Code-Review+1

Should we document that we expect the time on scheduler(s) and Zookeeper cluster to be reasonably in sync? If e.g. the clock of the scheduler is ahead more than 5min, we might start cleaning up partial event data nodes. Probably more of a theoretical concern though.

This is also a breaking change, in that it requires an update of all components that are using event queues at once. We might also leak old events from before the update that were not processed.

Patch-set: 4
Reviewer: Gerrit User 27582 <27582@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+1