Clean up leaked merge request params
Now that merge request parameters are sharded outside of a transaction,
we might write them but fail to write the actual merge request.

To address that, add a periodic check for leaked merge parameters to the
general cleanup thread in the Zuul scheduler. The same cleanup thread
also deals with leaked build parameters.

Change-Id: I29005816158e3d60a2bb0d6d46d3ff91e795e692