Use a new NodeRequest object when re-submitting node requests
We encountered an issue where Zuul:

* submitted a node request
* nodepool fulfilled it
* zuul received the ZK watch and refreshed the NodeRequest object
* zuul submitted the node provisioned event to the event queue
* ZK was disconnected/reconnected
* zuul processed the node provisioned event
* zuul found the node request no longer existed (because it's ephemeral)
* zuul resubmitted the node request

Because the NodeRequest object had the provisioned node information
attached to it, the re-submitted request was created with an
existing 'nodes' list.  Nodepool appended to that list and fulfilled
the new request (which requested 1 but received 2 nodes).  This caused
an exception in Zuul's nodepool request watch callback, which caused
Zuul to ignore that and all future updates to the node request.

To address this, we avoid re-using the NodeRequest object when
re-submitting a request, and ensure that we supply a new nodeset object
(with no nodes) to it.

The nodepool.requests dictionary was effectively only used for stats,
so that is replaced with a counter; it will need to be replaced again
as part of the multi-scheduler work.

Change-Id: I72f5ed7ad53e44d77b37870546daf61b8a4e7e09