Clear nodeset when re-submitting node requests
We encountered an issue where Zuul:

* submitted a node request
* nodepool fulfilled it
* zuul received the ZK watch and refreshed the NodeRequest object
* zuul submitted the node provisioned event to the event queue
* ZK was disconnected/reconnected
* zuul processed the node provisioned event
* zuul found the node request no longer existed (because it's ephemeral)
* zuul resubmitted the node request

Because the NodeRequest object had the provisioned node information
attached to it, the re-submitted request was created with an
existing 'nodes' list.  Nodepool appended to that list and fulfilled
the new request (which requested 1 but received 2 nodes).  This caused
an exception in Zuul's nodepool request watch callback, which caused
Zuul to ignore that and all future updates to the node request.

To address this, we make a new copy of the nodeset without any allocated
node info when re-submitting a request.

This contains an unrelated change to the event id handling from an earlier
revision; it is kept because it will simplify future changes which eliminate
the node request cache altogether.

Change-Id: I72f5ed7ad53e44d77b37870546daf61b8a4e7e09