Update Github SHA cache on PR update
It seems that the update of the SHA-to-PR cache in `getPull()` was
accidentally removed in Ib14bb387dbd233ff252cf57be7f0517770ade037.

This will lead to stale data in the cache and with that to events not
being processed, e.g. in case of an already resolved issue with multiple
PRs with the same HEAD SHA.

Change-Id: I178b63029d7a60523af4acefb3f05bff3daeb2c5