Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> I guess color isn't an option?  I think https://github.com/github/markup/issues/1440 says "no" but I'm not 100% sure that's applicable to the areas we're concerned with.

I had something completely different in mind but that would only work with colors. Now if you tell me that colors aren't working in GH flavored markdown that really is crap.

My original idea was to show the build results in a similar way like we do in the zuul UI, e.g. the build list on the buildset result page [1] just in form of a simple markdown list and not a table. That would also have brought the attention more to the colors (which could have been applied to the result text) and not to the emojis.

As this doesn't seem to be working, I think we have to use the icons/emojis to reflect the build result, but maybe we could do this with something less "fancy"? What I'm thinking about is something like the green checkmark icon that Travis CI shows on its build results [2] (Travis was the first thing that came into my mind when thinking about colored CI results in GitHub, but it looks like they use much less color than I thought). Maybe it's also worth to check out how other CI systems provide their results to GitHub.

[1]: https://zuul.opendev.org/t/opendev/buildset/c66fff1830a942cdbd4ed0581527f5a1
[2]: https://github.com/felixedel/flirror/pull/80/checks?check_run_id=1028135312

Patch-set: 3
CC: Gerrit User 27952 <27952@4a232e18-c5a9-48ee-94c0-e04e7cca6543>