DNM: review companion for merger/executor api refactor
The JobRequestQueue is a combination of the ExecutorApi and
the MergerApi.

To create it, I started with the MergerApi, genericized it, then
added the extra bits the ExecutorApi needs.

This is a synthetic change which shows the JobRequestQueue as a
diff against the original MergerApi.

(I also re-ordered some methods, so the final file in this change
is slightly different than the file in the refactor, but only in
the order of methods.  In this change, they match the original
order for ease of diffing.)

Change-Id: Iaf79776cf10b0124076ebb8f26d32329686d0a98